+++ rsa-pkcs1-sign.c
@@ -33,7 +33,7 @@
 
 int
 rsa_pkcs1_sign(const struct rsa_private_key *key,
-	       unsigned length, const uint8_t *digest_info,
+	       size_t length, const uint8_t *digest_info,
 	       mpz_t s)
 {  
   if (pkcs1_rsa_digest_encode (s, key->size, length, digest_info))
